feat(argo): adapt argo runtime to generate an exhaustive list of parameters and include default values
