+++ dagger/input/__init__.py
@@ -4,4 +4,5 @@ from dagger.input.empty_default_value import EmptyDefaultValue  # noqa
 from dagger.input.from_node_output import FromNodeOutput  # noqa
 from dagger.input.from_param import FromParam  # noqa
 from dagger.input.validators import split_required_and_optional_inputs  # noqa
+from dagger.input.validators import validate_and_clean_parameters  # noqa
 from dagger.input.validators import validate_name  # noqa
