+++ dagger/runtime/argo/workflow_spec.py
@@ -4,7 +4,12 @@ from typing import Any, Dict, List, Mapping, Sequence, Union
 
 from dagger.dag import DAG, Node
 from dagger.dag import SupportedInputs as SupportedDAGInputs
-from dagger.input import FromNodeOutput, FromParam, split_required_and_optional_inputs
+from dagger.input import (
+    FromNodeOutput,
+    FromParam,
+    split_required_and_optional_inputs,
+    validate_and_clean_parameters,
+)
 from dagger.runtime.argo.extra_spec_options import with_extra_spec_options
 from dagger.runtime.argo.workflow import Workflow
 from dagger.serializer import Serializer
@@ -38,7 +43,7 @@ def workflow_spec(
     ValueError
         If any of the extra_spec_options collides with a property used by the runtime.
     """
-    _validate_inputs(dag.inputs, workflow.params)
+    params = validate_and_clean_parameters(dag.inputs, workflow.params)
 
     spec = {
         "entrypoint": BASE_DAG_NAME,
@@ -46,12 +51,12 @@ def workflow_spec(
             node=dag,
             container_image=workflow.container_image,
             container_command=workflow.container_entrypoint_to_dag_cli,
-            params=workflow.params,
+            params=params,
         ),
     }
 
     if workflow.params:
-        spec["arguments"] = _workflow_spec_arguments(workflow.params)
+        spec["arguments"] = _workflow_spec_arguments(params)
 
     spec = with_extra_spec_options(
         original=spec,
@@ -62,20 +67,9 @@ def workflow_spec(
     return spec
 
 
-def _validate_inputs(
-    inputs: Mapping[str, Union[FromParam, FromNodeOutput]],
+def _workflow_spec_arguments(
     params: Mapping[str, Any],
-):
-    """Validate the workflow contains all the required parameters."""
-    required_inputs, _ = split_required_and_optional_inputs(inputs)
-    missing_params = required_inputs.keys() - params.keys()
-    if missing_params:
-        raise ValueError(
-            f"The following inputs are required for this workflow: {sorted(list(required_inputs))}. However, only the following parameters were actually supplied: {sorted(list(params))}. We are missing: {sorted(list(missing_params))}.",
-        )
-
-
-def _workflow_spec_arguments(params: Mapping[str, Any]) -> Mapping[str, Any]:
+) -> Mapping[str, Any]:
     return {
         "parameters": [
             {"name": param_name, "value": params[param_name]} for param_name in params
